issue #2050: unsigned int to int

+++ cocos/2d/CCTMXLayer.cpp
@@ -61,7 +61,7 @@ bool TMXLayer::initWithTilesetInfo(TMXTilesetInfo *tilesetInfo, TMXLayerInfo *la
         texture = Director::getInstance()->getTextureCache()->addImage(tilesetInfo->_sourceImage.c_str());
     }
 
-    if (SpriteBatchNode::initWithTexture(texture, (unsigned int)capacity))
+    if (SpriteBatchNode::initWithTexture(texture, static_cast<ssize_t>(capacity)))
     {
         // layerInfo
         _layerName = layerInfo->_name;
@@ -85,7 +85,7 @@ bool TMXLayer::initWithTilesetInfo(TMXTilesetInfo *tilesetInfo, TMXLayerInfo *la
         Point offset = this->calculateLayerOffset(layerInfo->_offset);
         this->setPosition(CC_POINT_PIXELS_TO_POINTS(offset));
 
-        _atlasIndexArray = ccCArrayNew((unsigned int)totalNumberOfTiles);
+        _atlasIndexArray = ccCArrayNew(totalNumberOfTiles);
 
         this->setContentSize(CC_SIZE_PIXELS_TO_POINTS(Size(_layerSize.width * _mapTileSize.width, _layerSize.height * _mapTileSize.height)));
 
@@ -161,12 +161,12 @@ void TMXLayer::setupTiles()
     // Parse cocos2d properties
     this->parseInternalProperties();
 
-    for (unsigned int y=0; y < _layerSize.height; y++) 
+    for (int y=0; y < _layerSize.height; y++)
     {
-        for (unsigned int x=0; x < _layerSize.width; x++) 
+        for (int x=0; x < _layerSize.width; x++)
         {
-            unsigned int pos = (unsigned int)(x + _layerSize.width * y);
-            unsigned int gid = _tiles[ pos ];
+            int pos = (int)(x + _layerSize.width * y);
+            int gid = _tiles[ pos ];
 
             // gid are stored in little endian.
             // if host is big endian, then swap
@@ -231,7 +231,7 @@ void TMXLayer::parseInternalProperties()
     }
 }
 
-void TMXLayer::setupTileSprite(Sprite* sprite, Point pos, unsigned int gid)
+void TMXLayer::setupTileSprite(Sprite* sprite, Point pos, int gid)
 {
     sprite->setPosition(getPositionAt(pos));
     sprite->setVertexZ((float)getVertexZForPos(pos));
@@ -252,7 +252,7 @@ void TMXLayer::setupTileSprite(Sprite* sprite, Point pos, unsigned int gid)
         sprite->setPosition(Point(getPositionAt(pos).x + sprite->getContentSize().height/2,
            getPositionAt(pos).y + sprite->getContentSize().width/2 ) );
 
-        unsigned int flag = gid & (kTMXTileHorizontalFlag | kTMXTileVerticalFlag );
+        int flag = gid & (kTMXTileHorizontalFlag | kTMXTileVerticalFlag );
 
         // handle the 4 diagonally flipped states.
         if (flag == kTMXTileHorizontalFlag)
@@ -319,7 +319,7 @@ Sprite * TMXLayer::getTileAt(const Point& pos)
     CCASSERT(_tiles && _atlasIndexArray, "TMXLayer: the tiles map has been released");
 
     Sprite *tile = nullptr;
-    unsigned int gid = this->getTileGIDAt(pos);
+    int gid = this->getTileGIDAt(pos);
 
     // if GID == 0, then no tile is present
     if (gid) 
@@ -348,14 +348,14 @@ Sprite * TMXLayer::getTileAt(const Point& pos)
     return tile;
 }
 
-unsigned int TMXLayer::getTileGIDAt(const Point& pos, ccTMXTileFlags* flags/* = nullptr*/)
+int TMXLayer::getTileGIDAt(const Point& pos, ccTMXTileFlags* flags/* = nullptr*/)
 {
     CCASSERT(pos.x < _layerSize.width && pos.y < _layerSize.height && pos.x >=0 && pos.y >=0, "TMXLayer: invalid position");
     CCASSERT(_tiles && _atlasIndexArray, "TMXLayer: the tiles map has been released");
 
-    int idx = (int)(pos.x + pos.y * _layerSize.width);
+    int idx = static_cast<int>((pos.x + pos.y * _layerSize.width));
     // Bits on the far end of the 32-bit global tile ID are used for tile flags
-    unsigned int tile = _tiles[idx];
+    int tile = _tiles[idx];
 
     // issue1264, flipped tiles can be changed dynamically
     if (flags) 
@@ -367,7 +367,7 @@ unsigned int TMXLayer::getTileGIDAt(const Point& pos, ccTMXTileFlags* flags/* = 
 }
 
 // TMXLayer - adding helper methods
-Sprite * TMXLayer::insertTileForGID(unsigned int gid, const Point& pos)
+Sprite * TMXLayer::insertTileForGID(int gid, const Point& pos)
 {
     Rect rect = _tileSet->rectForGID(gid);
     rect = CC_RECT_PIXELS_TO_POINTS(rect);
@@ -405,7 +405,7 @@ Sprite * TMXLayer::insertTileForGID(unsigned int gid, const Point& pos)
     return tile;
 }
 
-Sprite * TMXLayer::updateTileForGID(unsigned int gid, const Point& pos)    
+Sprite * TMXLayer::updateTileForGID(int gid, const Point& pos)
 {
     Rect rect = _tileSet->rectForGID(gid);
     rect = Rect(rect.origin.x / _contentScaleFactor, rect.origin.y / _contentScaleFactor, rect.size.width/ _contentScaleFactor, rect.size.height/ _contentScaleFactor);
@@ -427,7 +427,7 @@ Sprite * TMXLayer::updateTileForGID(unsigned int gid, const Point& pos)
 
 // used only when parsing the map. useless after the map was parsed
 // since lot's of assumptions are no longer true
-Sprite * TMXLayer::appendTileForGID(unsigned int gid, const Point& pos)
+Sprite * TMXLayer::appendTileForGID(int gid, const Point& pos)
 {
     Rect rect = _tileSet->rectForGID(gid);
     rect = CC_RECT_PIXELS_TO_POINTS(rect);
@@ -458,7 +458,7 @@ static inline int compareInts(const void * a, const void * b)
     return ((*(int*)a) - (*(int*)b));
 }
 
-ssize_t TMXLayer::atlasIndexForExistantZ(unsigned int z)
+ssize_t TMXLayer::atlasIndexForExistantZ(int z)
 {
     int key=z;
     int *item = (int*)bsearch((void*)&key, (void*)&_atlasIndexArray->arr[0], _atlasIndexArray->num, sizeof(void*), compareInts);
@@ -486,23 +486,23 @@ ssize_t TMXLayer::atlasIndexForNewZ(int z)
 }
 
 // TMXLayer - adding / remove tiles
-void TMXLayer::setTileGID(unsigned int gid, const Point& pos)
+void TMXLayer::setTileGID(int gid, const Point& pos)
 {
     setTileGID(gid, pos, (ccTMXTileFlags)0);
 }
 
-void TMXLayer::setTileGID(unsigned int gid, const Point& pos, ccTMXTileFlags flags)
+void TMXLayer::setTileGID(int gid, const Point& pos, ccTMXTileFlags flags)
 {
     CCASSERT(pos.x < _layerSize.width && pos.y < _layerSize.height && pos.x >=0 && pos.y >=0, "TMXLayer: invalid position");
     CCASSERT(_tiles && _atlasIndexArray, "TMXLayer: the tiles map has been released");
     CCASSERT(gid == 0 || gid >= _tileSet->_firstGid, "TMXLayer: invalid gid" );
 
     ccTMXTileFlags currentFlags;
-    unsigned int currentGID = getTileGIDAt(pos, &currentFlags);
+    int currentGID = getTileGIDAt(pos, &currentFlags);
 
     if (currentGID != gid || currentFlags != flags) 
     {
-        unsigned gidAndFlags = gid | flags;
+        int gidAndFlags = gid | flags;
 
         // setting gid=0 is equal to remove the tile
         if (gid == 0)
@@ -517,7 +517,7 @@ void TMXLayer::setTileGID(unsigned int gid, const Point& pos, ccTMXTileFlags fla
         // modifying an existing tile with a non-empty tile
         else 
         {
-            unsigned int z = (unsigned int)(pos.x + pos.y * _layerSize.width);
+            int z = pos.x + pos.y * _layerSize.width;
             Sprite *sprite = static_cast<Sprite*>(getChildByTag(z));
             if (sprite)
             {
@@ -570,11 +570,11 @@ void TMXLayer::removeTileAt(const Point& pos)
     CCASSERT(pos.x < _layerSize.width && pos.y < _layerSize.height && pos.x >=0 && pos.y >=0, "TMXLayer: invalid position");
     CCASSERT(_tiles && _atlasIndexArray, "TMXLayer: the tiles map has been released");
 
-    unsigned int gid = getTileGIDAt(pos);
+    int gid = getTileGIDAt(pos);
 
     if (gid) 
     {
-        unsigned int z = (unsigned int)(pos.x + pos.y * _layerSize.width);
+        int z = pos.x + pos.y * _layerSize.width;
         ssize_t atlasIndex = atlasIndexForExistantZ(z);
 
         // remove tile from GID map
@@ -676,17 +676,17 @@ Point TMXLayer::getPositionForHexAt(const Point& pos)
 int TMXLayer::getVertexZForPos(const Point& pos)
 {
     int ret = 0;
-    unsigned int maxVal = 0;
+    int maxVal = 0;
     if (_useAutomaticVertexZ)
     {
         switch (_layerOrientation) 
         {
         case TMXOrientationIso:
-            maxVal = (unsigned int)(_layerSize.width + _layerSize.height);
-            ret = (int)(-(maxVal - (pos.x + pos.y)));
+            maxVal = static_cast<int>(_layerSize.width + _layerSize.height);
+            ret = static_cast<int>(-(maxVal - (pos.x + pos.y)));
             break;
         case TMXOrientationOrtho:
-            ret = (int)(-(_layerSize.height-pos.y));
+            ret = static_cast<int>(-(_layerSize.height-pos.y));
             break;
         case TMXOrientationHex:
             CCASSERT(0, "TMX Hexa zOrder not supported");
